shell: Split old custom functions in zsh function or bash scripts
deleted: Mostly outdated, in need of substancial rework or not useful
anymore.

- brewrelink: More complex operation, potentially time-consuming
- dash: Simple application launcher
- watchme: Debugging utility [HF: brew install watch]

bash scripts: Involves more complex operations that might be 
time-consuming and is less likely to be used in every shell session. 
They benefit from being a custom script that can be called when needed.

- [HF: None yet]

zsh autoload functions: Are relatively short, specific to the shell
environment, and likely to be used frequently in interactive sessions. 
They benefit from being quickly accessible in memory.

- ask_if_empty: Frequently used, shell-specific utility
- clean-resource-forks: macOS-specific utility
- close: Simple OS interaction
- find-resource-forks: macOS-specific utility
- fix-zip: File manipulation utility
- git-anonymise-repo: Git utility
- git-commit-dates: Git utility
- isadmin: Simple system check
- man: macOS-specific utility override
- open-site: Development workflow utility
- port2pid: System utility
- safari_history: macOS-specific utility
- source_from_brew: Brew-specific utility
- source_from_path: General utility
- sudoadd, sudorm, sudoaupdate: System administration utilities
- super: System administration utility
- week-days-en, week-days-pt: Date formatting utilities